# docs/openapi.json
{"openapi":"3.0.3","info":{"title":"Secondhand MCP REST API","description":"Self-hosted HTTP front-end over the marketplace abstraction. Searches Facebook Marketplace, eBay, Depop and Poshmark through the same registry the MCP server uses. Interactive docs below; raw spec at /docs/json.","version":"1.0.0"},"components":{"schemas":{}},"paths":{"/health":{"get":{"summary":"Service health","tags":["health"],"description":"Reports that the API is up and lists the marketplaces currently registered (eBay/Depop/Poshmark may be absent if their credentials/browser are unavailable).","responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string"},"server":{"type":"string"},"marketplaces":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"displayName":{"type":"string"},"requiresAuth":{"type":"boolean"}}}}}}}}}}}},"/v1/locations/resolve":{"get":{"summary":"Resolve a place name to coordinates","tags":["locations"],"description":"Asks a marketplace to resolve a human-readable place to latitude/longitude. For Facebook the server reads Facebook's own location search (it is not hard-coded); the input and resolved name/coords are returned token-for-token.","parameters":[{"schema":{"type":"string","default":"facebook"},"in":"query","name":"marketplace","required":false,"description":"Marketplace id. Valid values: facebook, ebay, depop, poshmark"},{"schema":{"type":"string"},"in":"query","name":"location","required":true,"description":"Place to resolve, e.g. \"Santo Domingo, Dominican Republic\""}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"input":{"type":"string"},"name":{"type":"string"},"latitude":{"type":"number"},"longitude":{"type":"number"}}}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"input":{"type":"string"},"found":{"type":"boolean"},"error":{"type":"string"}}}}}},"501":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/v1/search":{"post":{"summary":"Search live listings on a marketplace","tags":["search"],"description":"Searches a marketplace using the requested query and filters. The response includes the resolved search location (the coordinates the adapter actually searched with) alongside the normalized listings. minPrice/maxPrice are applied by the target marketplace in its own currency.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["query"],"properties":{"marketplace":{"type":"string","description":"Marketplace id. Valid values: facebook, ebay, depop, poshmark","default":"facebook"},"query":{"type":"string","description":"Search terms, e.g. \"iphone 15\""},"location":{"type":"string","description":"Place to search around (Facebook only)"},"radius":{"type":"number","description":"Search radius in miles (alias; Facebook only)"},"radiusMiles":{"type":"number","description":"Search radius in miles (Facebook only)"},"minPrice":{"type":"number","minimum":0},"maxPrice":{"type":"number","minimum":0},"limit":{"type":"number","minimum":1},"offset":{"type":"number","minimum":0},"showSold":{"type":"boolean"},"sort":{"type":"string","description":"Sort order (Depop, Poshmark)"},"condition":{"type":"string"},"category":{"type":"string"},"brand":{"type":"string"},"department":{"type":"string"},"sizes":{"type":"array","items":{"type":"string"}},"colors":{"type":"array","items":{"type":"string"}}}}}}},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"search":{"type":"object","properties":{"marketplace":{"type":"string"},"query":{"type":"string"},"location":{"type":"object","properties":{"input":{"type":"string"},"name":{"type":"string"},"latitude":{"type":"number"},"longitude":{"type":"number"}}},"radiusMiles":{"type":"number"},"minPrice":{"type":["null","number"]},"maxPrice":{"type":["null","number"]}}},"success":{"type":"boolean"},"marketplace":{"type":"string"},"listings":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"price":{"type":"string"},"priceNumeric":{"type":"number"},"currency":{"type":"string"},"location":{"type":"string"},"description":{"type":"string"},"url":{"type":"string"},"images":{"type":"array","items":{"type":"string"}},"seller":{"type":"string"},"condition":{"type":"string"},"marketplace":{"type":"string"},"scrapedAt":{"type":"string"}}}},"totalFound":{"type":"number"},"note":{"type":"string"},"error":{"type":"string"}}}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"502":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"search":{"type":"object"},"success":{"type":"boolean"},"listings":{"type":"array"},"error":{"type":"string"}}}}}}}}},"/v1/listings/{marketplace}/{id}":{"get":{"summary":"Full details for a single listing","tags":["listings"],"description":"Returns the normalized listing detail object for a marketplace id from a search result: description, every photo, location, seller and delivery options.","parameters":[{"schema":{"type":"string"},"in":"path","name":"marketplace","required":true,"description":"Marketplace id. Valid values: facebook, ebay, depop, poshmark"},{"schema":{"type":"string"},"in":"path","name":"id","required":true,"description":"Listing id from a search result"}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"marketplace":{"type":"string"},"id":{"type":"string"},"description":{"type":"string"},"images":{"type":"array","items":{"type":"string"}},"location":{"type":"string"},"locationCoords":{"type":"object","properties":{"latitude":{"type":"number"},"longitude":{"type":"number"}}},"seller":{"type":"string"},"deliveryTypes":{"type":"array","items":{"type":"string"}},"isShippingOffered":{"type":"boolean"},"url":{"type":"string"}}}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"marketplace":{"type":"string"},"id":{"type":"string"},"error":{"type":"string"}}}}}},"501":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}}},"tags":[{"name":"health","description":"Service and marketplace status"},{"name":"locations","description":"Resolve place names to coordinates"},{"name":"search","description":"Search live listings on a marketplace"},{"name":"listings","description":"Fetch full details for one listing"}]}